move isDevServerRunning to metro-service
