+++ packages/metro-service/package.json
@@ -26,7 +26,8 @@
   "dependencies": {
     "@rnx-kit/console": "^1.0.0",
     "@rnx-kit/tools-language": "^2.0.0",
-    "chalk": "^4.1.0"
+    "chalk": "^4.1.0",
+    "node-fetch": "^2.6.7"
   },
   "peerDependencies": {
     "@office-iss/react-native-win32": "*",
@@ -54,6 +55,7 @@
     "@react-native-community/cli-types": "^11.0.0",
     "@rnx-kit/scripts": "*",
     "@types/node": "^18.0.0",
+    "@types/node-fetch": "^2.6.5",
     "eslint": "^8.0.0",
     "jest": "^29.2.1",
     "metro": "^0.76.5",
