+++ packages/metro-service/src/index.ts
@@ -3,6 +3,6 @@ export type { BundleArgs } from "./bundle";
 export { loadMetroConfig } from "./config";
 export type { MetroConfigOverrides } from "./config";
 export { ramBundle } from "./ramBundle";
-export { startServer } from "./server";
+export { isDevServerRunning, startServer } from "./server";
 export { createTerminal } from "./terminal";
 export type { MetroTerminal } from "./terminal";
